requirements.txt
watchfiles
torch
transformers
Pillow
timm